classification-inference-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
onnxruntime==1.16.0
pillow==10.0.1
numpy==1.24.3
pydantic==2.4.2
prometheus-client==0.12.0
starlette-prometheus==0.9.0